Moved glue.qt.qtutil.ComponentIDCombo to glue.core.qt.component_id_combo.ComponentIDCombo
