+++ CHANGES.md
@@ -91,6 +91,7 @@ v0.7 (unreleased)
   ``glue.qt.widgets.settings_editor``                | ``glue.app.qt.settings_editor``
   ``glue.qt.qtutil.data_wizard``                     | ``glue.dialogs.data_wizard.qt.data_wizard``
   ``glue.qt.mime``                                   | ``glue.core.qt.mime``
+  ``glue.qt.qtutil.ComponentIDCombo``                | ``glue.core.qt.component_id_combo``
 
 * Improved under-the-hood creation of ROIs for Scatter and Histogram Clients. [#676]
